Update README.txt
The guide in https://apps.developer.homey.app/app-store/guidelines
ask to keep the description short and concise, without any markdown
or other formats.

We ideally shouldn't include any links either, but as long as the
pairing process doesn't have a login flow that automatically fetches
the secrets and device IDs we need to tell the user how to get started.